Refactorisation : creer un dossier metrics, sortir les exports dans un fichiers disctincts par produit
