+++ src/handlers/authentication.rs
@@ -12,6 +12,7 @@ use actix_web::{
 use futures::future::{ok, LocalBoxFuture, Ready};
 use log::{debug, warn};
 
+
 /// Middleware factory for authentication
 pub struct Authentication;
 
@@ -52,7 +53,7 @@ where
         let authenticate_result = authenticate(&req);
 
         // Record authentication attempts - we'll use app_data to get metrics if available
-        if let Some(metrics) = req.app_data::<actix_web::web::Data<crate::metrics::Metrics>>() {
+        if let Some(metrics) = req.app_data::<actix_web::web::Data<crate::metrics::metrics::Metrics>>() {
             let metrics = metrics.clone();
 
             if authenticate_result.is_err() {
