+++ src/lib.rs
@@ -7,13 +7,15 @@ pub mod config;
 pub mod error;
 pub mod file_utils;
 pub mod handlers;
+pub mod metrics;
 pub mod models;
 pub mod queue_manager;
-pub mod metrics;
 
 // Re-export common types for easier access
 pub use config::HandlerConfig;
+pub use config::MetricsConfig;
 pub use error::HandlerError;
 pub use handlers::{cancel_transcription, transcribe, transcription_result, transcription_status};
-pub use models::{TranscriptionResponse, ErrorResponse, SuccessResponse};
+pub use metrics::metrics::Metrics;
+pub use models::{ErrorResponse, SuccessResponse, TranscriptionResponse};
 pub use queue_manager::{QueueManager, TranscriptionJob, WhisperConfig};
