+++ src/metrics/mod.rs
@@ -1,0 +1,9 @@
+// Whisper API metrics for monitoring and observability
+//
+// This module contains the metrics for the Whisper API.
+// It provides functionality to track and report various metrics
+
+pub mod metrics;
+pub mod null;
+pub mod prometheus;
+pub mod statsd;
