+++ src/metrics/statsd.rs
@@ -1,0 +1,41 @@
+use crate::metrics::metrics::MetricsExporter;
+
+use async_trait::async_trait;
+use log::debug;
+
+/// TODO : StatsD implementation (placeholder for future implementation)
+pub struct StatsDExporter {
+    // Future implementation for StatsD
+    _endpoint: String,
+}
+
+impl StatsDExporter {
+    pub fn new(endpoint: String) -> Self {
+        Self {
+            _endpoint: endpoint,
+        }
+    }
+}
+
+#[async_trait]
+impl MetricsExporter for StatsDExporter {
+    async fn increment(&self, _name: &str, _labels: &[(&str, &str)]) {
+        // TODO: implement StatsD counter increment
+        debug!("StatsD increment not yet implemented");
+    }
+
+    async fn set_gauge(&self, _name: &str, _value: f64, _labels: &[(&str, &str)]) {
+        // TODO: implement StatsD gauge set
+        debug!("StatsD gauge set not yet implemented");
+    }
+
+    async fn observe_histogram(&self, _name: &str, _value: f64, _labels: &[(&str, &str)]) {
+        // TODO: implement StatsD histogram observe
+        debug!("StatsD histogram observe not yet implemented");
+    }
+
+    async fn export(&self) -> Result<Vec<u8>, String> {
+        // StatsD doesn't export, it pushes metrics
+        Ok(vec![])
+    }
+}
